# docs/api_swagger.yaml
openapi: 3.0.3
info:
  title: Projeto monitoramento em transporte de pets - OpenAPI 3.0
  description: |-
      Esta é a documentação da API para comunicação com o backend do projeto
      
      **Esta API requer autenticação com JWT**
      

       
  version: v1
servers:
  - url: http://localhost/v1/
  - url: http://192.168.0.104:3333/v1/
...
paths:
...
  /pet/{petId}:
    get:
      tags:
        - pet
      summary: Recupera um pet
      description: Retorna um animal
      operationId: getPetById
      parameters:
        - name: petId
          in: path
          description: ID do 
          
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetGET'          

        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: []
        
    put:
      tags:
        - pet
      summary: Atualiza os dados do pet
      description: ''
      operationId: updatePetWithForm
      parameters:
        - name: petId
          in: path
          description: ID do pet que deseja-se atualizar
          required: true
          schema:
            type: string
      requestBody:
        description: Atualiza os dados de um pet existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
        required: true
      responses:
        '201':
          description: Atualizado com sucesso
        '405':
          description: Entrada Invalida
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: []
        
    delete:
      tags:
        - pet
      summary: Remover pet indicado
      description: Remove o registro do pet
      operationId: deletePet
      parameters:
        
        - name: petId
          in: path
          description: Id do pet a ser removido
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Pet não encontrado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: []
...
  /trip:
    post:
      tags:
        - Viagens
      summary: Cadastra uma viagem para o usuário autenticado
      requestBody:
        description: Cadastrar Viagem
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viagem'
      responses:
        '201':
          description: Cadastrado com sucesso

        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: []
    get:
      tags:
        - Viagens
      summary: Busca todas as viagens do usuário
      
      responses:
        '200':
            $ref: '#/components/responses/ViagemArray'          

        '404':
          description: Userario não encontrado
          
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: []
  /trip/pet/{petID}:
    get:
      tags:
        - Viagens
      summary: Busca as viagens do pet selecionado
      parameters:
        - name: petID
          in: path
          description: Id do pet
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViagemGET' 
  /trip/{tripId}:
    get:
      tags:
        - Viagens
      summary: Busca a viagem selecionada
      parameters:
        - name: tripId
          in: path
          description: Id da viagem
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViagemGET'
                
        '404':
          description: Viagem não encontrada
          
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Viagens
      summary: Apaga a viagem selecionada
      parameters:
        - name: tripId
          in: path
          description: Id da viagem
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removido com sucesso

        '404':
          description: Viagem não encontrada
          
        '401':
          $ref: '#/components/responses/UnauthorizedError'
                  
    put:
      tags:
        - Viagens
      summary: Atualiza a viagem desejada
      parameters:
        - name: tripId
          in: path
          description: Id da viagem
          required: true
          schema:
            type: string
      requestBody:
        description: Atualizar viagem
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viagem'
      responses:
        '201':
          description: Cadastrado com sucesso

        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
      security:
        - bearerAuth: [] 
...
components:
  schemas:
...
    Pet:
      required:
        - id
        - name
      type: object
      properties:
        name:
          type: string
          example: doggie
        breed:
          type: string
          example: Caramelo
        category:
          type: string
          example: Medio Porte
        age:
          type: string
          example: 1
...
    PetGET:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: doggie
        breed:
          type: string
          example: Caramelo
        category:
          type: string
          example: Medio Porte
...
    Viagem:
      type: object
      properties:
        petId:
          type: string
          example: 7dbb95a2-3489-11ed-a261-0242ac120002
        caixaId: 
          type: string
          example: 7dbb95a2-3489-11ed-a274-0242ac120002
        date:
          type: string
          example: "2022-08-24 09:10:00"
    ViagemGET:
      type: object
      properties:
        id:
          type: string
          example: 7dbb95a2-3489-11ed-a261-0242ac120002
        petId:
          type: string
          example: 7dbb95a2-3489-11ed-a261-0242ac120002
        caixaID: 
          type: string
          example: 7dbb95a2-3489-11ed-a274-0242ac120002
        date:
          type: string
          example: "2022-08-24 09:10:00"
...
  responses:
    UnauthorizedError:
      description: Operação não autorizada, verifique suas permissões
    ViagemArray:
      description: Lista de viagens
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ViagemGET'
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes